preliminary win32 port
git-svn-id: https://svn.osgeo.org/gdal@16780 f0d54148-0727-0410-94bb-9a71ac55c965

+++ sandbox/warmerdam/pcidsk/src/cbandinterleavedchannel.cpp
@@ -125,7 +125,7 @@ int CBandInterleavedChannel::ReadBlock( int block_index, void *buffer )
     else
     {
         int  i;
-        PCIDSKBuffer line_from_disk( pixel_offset * (width-1) + pixel_size );
+        PCIDSKBuffer line_from_disk( (int) (pixel_offset*(width-1) + pixel_size) );
         char *this_pixel;
 
         MutexHolder holder( *io_mutex_p );
